move page and update out of elements folder
